% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.flxTrans}
\alias{plot.flxTrans}
\title{plot}
\usage{
\method{plot}{flxTrans}(x, ...)
}
\arguments{
\item{x}{flexTrans object}

\item{...}{Additional arguments passed to the plot function}
}
\description{
Plots a flextrans object
}
\examples{
d1 <- rnorm(100)
d2 <- rgamma(1000,1,5)
l <- list(d1, d2)
obj <- trans(l)
plot(obj)
}
