% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{trans}
\usage{
trans(l, n_theta = 3)
}
\arguments{
\item{l}{A vector or list of vectors with that will be transformed to a gaussian distribution.}

\item{n_theta}{A integer specifying the number of theta parameters to use in the transformation.
If a vector of integers is supplied, there will be one transformation per element.}
}
\value{
Returns an object containing a data frame with the transformed values, a data frame with predictions and the original
vector or list of vectors. Use \code{\link[=plot]{plot()}} on this object to plot diagnostics.
}
\description{
Transforms a vector or a list of vectors to a gaussian distribution using a Taylor-like transformation function.
}
\examples{
x1 <- rgamma(100, 1, 5)
x2 <- rgamma(1000, 1, 5)
x3 <- rbeta(100, 2, 4)
l <- list(x1, x2, x3)
trans(l, n_theta = 3)
}
