% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantflex-package.R
\docType{package}
\name{quantflex-package}
\alias{quantflex}
\alias{quantflex-package}
\title{quantflex: quantflex}
\description{
Estimation of conditional and unconditional quantiles and conditional trajectory quantiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/martinthorsson/quantflex}
  \item Report bugs at \url{https://github.com/martinthorsson/quantflex/issues}
}

}
\author{
\strong{Maintainer}: Matteo Bottai \email{matteo.bottai@ki.se} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Martin Thorsson \email{martin.thorsson@gu.se} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
