% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{trans}
\usage{
trans(l, n_theta = 3)
}
\arguments{
\item{l}{A list with vectors or a vector to transform}

\item{n_theta}{A integer specifying the number of theta parameters to use in the transformation.
If a vector of integers is supplied, there will be one transformation per element.}
}
\value{
Returns a data frame with the transformed vectors in long format.
}
\description{
Transforms a vector or a list of vectors to a gaussian distribution using a Taylor-like transformation function.
}
\examples{
x1 <- rgamma(100, 1, 5)
x2 <- rgamma(1000, 1, 5)
x3 <- rbeta(100, 2, 4)
l <- list(x1, x2, x3)
trans(l, n_theta = 3)
}
